//#EJC: Versión 4.1.8  Corrección en fechas para carga y envío de datos.

+++ app/src/main/java/com/dtsgt/base/clsClasses.java
@@ -469,7 +469,7 @@ public class clsClasses {
     }
 
     public class clsP_cliente {
-        public int  codigo_cliente;
+        public long  codigo_cliente;
         public String codigo;
         public int  empresa;
         public String nombre;
