+++ app/src/main/java/com/dtsgt/mpos/Venta.java
@@ -3484,7 +3484,7 @@ public class Venta extends PBase {
     private double totalCredito() {
         Cursor dt;
         double cred=0;
-        int mm,yy;
+        long mm,yy;
         long ff;
 
         try {
